+++ examples/light.rs
@@ -13,7 +13,6 @@ use core::pin::pin;
 use log::info;
 
 use rs_matter_stack::ble::BluerGattPeripheral;
-use rs_matter_stack::endpoints::FnMatcher;
 use rs_matter_stack::matter::crypto::{default_crypto, Crypto};
 use rs_matter_stack::matter::dm::clusters::app::on_off;
 use rs_matter_stack::matter::dm::clusters::app::on_off::test::TestOnOffDeviceLogic;
@@ -87,18 +86,18 @@ fn main() -> Result<(), Error> {
         // Diagnostics) on top, because only it knows the network driver state.
         // Chain any extra Endpoint 0 clusters of your own the same way.
         .chain(
-            FnMatcher(|e, _| e == ROOT_ENDPOINT_ID),
+            |e, _| e == ROOT_ENDPOINT_ID,
             Async(WifiMatterStack::<0, ()>::root_handler(&(), &mut rand)),
         )
         // Our on-off cluster, on Endpoint 1
         .chain(
-            FnMatcher(|e, c| e == LIGHT_ENDPOINT_ID && c == TestOnOffDeviceLogic::CLUSTER.id),
+            |e, c| e == LIGHT_ENDPOINT_ID && c == TestOnOffDeviceLogic::CLUSTER.id,
             on_off::HandlerAsyncAdaptor(&on_off),
         )
         // Each Endpoint needs a Descriptor cluster too
         // Just use the one that `rs-matter` provides out of the box
         .chain(
-            FnMatcher(|e, c| e == LIGHT_ENDPOINT_ID && c == DescHandler::CLUSTER.id),
+            |e, c| e == LIGHT_ENDPOINT_ID && c == DescHandler::CLUSTER.id,
             Async(DescHandler::new(Dataver::new_rand(&mut rand)).adapt()),
         );